requirements.txt
langchain
openai
python-dotenv
google-search-results
langchain_community
chromadb
faiss-cpu
streamlit
streamlit-authenticator
bcrypt
